move hashindex to borg.algorithms, fixes #2586
