+++ src/borg/algorithms/_hashindex.c
