+++ src/borg/algorithms/hashindex.pyx
