+++ src/borg/repository.py
@@ -12,7 +12,6 @@ from itertools import islice
 import msgpack
 
 from .constants import *  # NOQA
-from .hashindex import NSIndex
 from .helpers import Error, ErrorWithTraceback, IntegrityError, format_file_size, parse_file_size
 from .helpers import Location
 from .helpers import ProgressIndicatorPercent
@@ -23,6 +22,7 @@ from .locking import Lock, LockError, LockErrorT
 from .logger import create_logger
 from .lrucache import LRUCache
 from .platform import SaveFile, SyncFile, sync_dir, safe_fadvise
+from .algorithms.hashindex import NSIndex
 from .algorithms.checksums import crc32
 from .crypto.file_integrity import IntegrityCheckedFile, FileIntegrityError
